+++ lc-lib/transports/tcp/transporttcpfactory.go
@@ -92,7 +92,7 @@ func (f *TransportTCPFactory) Defaults() {
 
 // NewTransport returns a new Transport interface using the settings from the
 // TransportTCPFactory.
-func (f *TransportTCPFactory) NewTransport(ctx context.Context, pool *addresspool.Pool, eventChan chan<- transports.Event, finishOnFail bool) transports.Transport {
+func (f *TransportTCPFactory) NewTransport(ctx context.Context, pool *addresspool.Pool, eventChan chan<- transports.Event) transports.Transport {
 	cancelCtx, shutdownFunc := context.WithCancel(ctx)
 
 	ret := &transportTCP{
@@ -100,7 +100,6 @@ func (f *TransportTCPFactory) NewTransport(ctx context.Context, pool *addresspoo
 		shutdownFunc: shutdownFunc,
 		config:       f,
 		netConfig:    transports.FetchConfig(f.config),
-		finishOnFail: finishOnFail,
 		pool:         pool,
 		eventChan:    eventChan,
 		backoff:      core.NewExpBackoff(pool.Server()+" Reconnect", f.Reconnect, f.ReconnectMax),